+++ packages/@dcl/ecs/src/components/extended/Material.ts
@@ -1,4 +1,4 @@
-import { ComponentDefinition, Entity, IEngine } from '../../engine'
+import { LastWriteWinElementSetComponentDefinition, Entity, IEngine } from '../../engine'
 import { Material, PBMaterial, PBMaterial_PbrMaterial, PBMaterial_UnlitMaterial } from '../generated/index.gen'
 import { AvatarTexture, Texture, TextureUnion, VideoTexture } from '../generated/types.gen'
 
@@ -25,7 +25,7 @@ export interface TextureHelper {
 /**
  * @public
  */
-export interface MaterialComponentDefinitionExtended extends ComponentDefinition<PBMaterial> {
+export interface MaterialComponentDefinitionExtended extends LastWriteWinElementSetComponentDefinition<PBMaterial> {
   /**
    * Texture helpers with constructor
    */
