+++ packages/@dcl/ecs/src/engine/index.ts
@@ -6,7 +6,8 @@ import { ISchema } from '../schemas/ISchema'
 import { MapResult, Spec } from '../schemas/Map'
 import { ByteBuffer } from '../serialization/ByteBuffer'
 import { crdtSceneSystem, OnChangeFunction } from '../systems/crdt'
-import { ComponentDefinition, createComponentDefinitionFromSchema } from './component'
+import { ComponentDefinition } from './component'
+import { createComponentDefinitionFromSchema } from './lww-element-set-component-definition'
 import { Entity, EntityContainer } from './entity'
 import { ReadonlyComponentSchema } from './readonly'
 import { SystemItem, SystemContainer, SystemFn, SYSTEMS_REGULAR_PRIORITY } from './systems'
@@ -14,7 +15,7 @@ import type { IEngine, IEngineOptions, MapComponentDefinition, PreEngine } from 
 export * from './input'
 export * from './readonly'
 export * from './types'
-export { Entity, ByteBuffer, ComponentDefinition, SystemItem, OnChangeFunction }
+export { Entity, ByteBuffer, SystemItem, OnChangeFunction }
 
 function preEngine(): PreEngine {
   const entityContainer = EntityContainer()
@@ -78,11 +79,11 @@ function preEngine(): PreEngine {
     const prev = componentsDefinition.get(componentId)
     if (prev) {
       // TODO: assert spec === prev.spec
-      return prev as ComponentDefinition<T>
+      return prev as components.LastWriteWinElementSetComponentDefinition<T>
     }
     const newComponent = createComponentDefinitionFromSchema<T>(componentName, componentId, schema)
     componentsDefinition.set(componentId, newComponent)
-    return newComponent as ComponentDefinition<T>
+    return newComponent as components.LastWriteWinElementSetComponentDefinition<T>
   }
 
   function defineComponent<T extends Spec>(
